+++ bcms/api.py
@@ -1,11 +1,11 @@
 import logging
 from typing import Union
 from dataclasses import dataclass
+from px_python_shared.well_known import ApplicationsWellKnown
 import requests
 
 from px_python_shared import (
     get_well_known_by_identifier,
-    get_well_known,
     make_bearer_headers,
     add_scheme_from_auth_host,
 )
@@ -40,81 +40,71 @@ class IotDeviceCreateResponse:
 
 
 class BackendAPI:
-    is_loaded: bool
-
     auth_host: Union[None, str]
     app_host: Union[None, str]
     identifier: Union[None, str]
-    device_id: Union[None, str]
 
     well_known: Union[None, dict]
 
-    def __init__(self) -> None:
-        self.is_loaded = False
-
+    def __init__(self, identifier: Union[str, None]) -> None:
         self.auth_host = None
         self.app_host = None
-        self.identifier = None
-        self.device_id = None
-
-        self.well_known = None
-
-    def load(self, identifier: Union[str, None] = None):
-        from px_device_identity import is_superuser_or_quit
-
-        is_superuser_or_quit()
-
-        log.info("Loading API with identifier %s ...", identifier)
-
+        # no identifier: offline use
         self.identifier = identifier
         if identifier is None:
             log.warning("Well known identifier is undefined.")
         else:
-            result = self.refresh_well_known()
-            if result:
-                self.is_loaded = True
+            log.info("Loading API with identifier %s ...", identifier)
+
+        self.well_known = None
+            
+    def ready_api(self) -> None:
+        """Check if identifier is set and refresh well known if necessary"""
+        from px_device_identity import is_superuser_or_quit
+        is_superuser_or_quit()
+        
+        if not self.identifier:
+            raise ValueError("Well known identifier is not set")
+        
+        if self.well_known is None:
+            self.refresh_well_known()
+    
+    def access_token_and_host(self) -> tuple:
+        """Get access token and host"""
+        from px_device_identity import Device
+
+        device = Device()
+        if not device.is_initiated:
+            raise ValueError("Device is not initiated: Has it been registered yet?")
+        
+        result = device.get_access_token()
+
+        return result["access_token"], device.properties.host
 
     def access_token(self) -> str:
         """Get access token"""
+        access_token, idp_host = self.access_token_and_host()
 
-        from px_device_identity import Device
+        return access_token
 
-        device = Device()
-        result = device.get_access_token()
-
-        return result["access_token"]
-
-    def refresh_well_known(self):
+    def refresh_well_known(self) -> ApplicationsWellKnown:
         """Refresh well known"""
-        from px_device_identity import Device
-
-        device = Device()
-        try:
-            well_known_list = get_well_known(device)
-            if "data" not in well_known_list:
-                log.warning("No well known found. API Error.")
-                return None
-
-            well_known = get_well_known_by_identifier(
-                well_known=well_known_list["data"], identifier=self.identifier
-            )
-
-            if well_known is None or "data" not in well_known:
-                log.warning("No well known found for %s.", self.identifier)
-                return None
-
-            self.app_host = add_scheme_from_auth_host(
-                well_known["data"].hostname, device.properties.host
-            )
-            return well_known
-        except Exception as err:
-            log.error("Error getting well known: %s", err)
-            return None
+        access_token, idp_host = self.access_token_and_host()
+        
+        self.well_known = get_well_known_by_identifier(
+            idp_host, access_token, self.identifier
+        )
+        
+        self.app_host = add_scheme_from_auth_host(
+            self.well_known.hostname, idp_host
+        )
+        
+        return self.well_known
 
     async def submit_iot_data(self, data: list):
         """Submit iot data to server"""
-
         log.debug("Submitting iot data %s", data)
+        self.ready_api()
 
         url = f"{self.app_host}/api/iot-devices/data/submit"
         res = requests.post(
@@ -127,8 +117,8 @@ class BackendAPI:
 
     def submit_iot_data_sync(self, data: list):
         """Submit iot data to server"""
-
         log.debug("Submitting iot data %s", data)
+        self.ready_api()
 
         url = f"{self.app_host}/api/iot-devices/data/submit"
         res = requests.post(
@@ -141,8 +131,8 @@ class BackendAPI:
 
     def iot_device_exists(self, address: str):
         """Check if iot device exists"""
-
         log.debug("Checking if iot device exists %s", address)
+        self.ready_api()
 
         url = f"{self.app_host}/api/iot-devices/exists"
         res = requests.post(
@@ -154,8 +144,8 @@ class BackendAPI:
         res.raise_for_status()
         data = res.json()
 
-        exists = "exists" in data and data["exists"] == True
-        remember = "rememberDevice" in data and data["rememberDevice"] == True
+        exists = "exists" in data and data["exists"] is True
+        remember = "rememberDevice" in data and data["rememberDevice"] is True
 
         if exists and remember:
             # allowMultipleDeviceRelationships=true
@@ -179,8 +169,8 @@ class BackendAPI:
         supported_data_types: Union[list, None] = None,
     ):
         """Create iot device"""
-
         log.info("Creating iot device %s", hardware_identifier)
+        self.ready_api()
 
         url = f"{self.app_host}/api/iot-devices"
         data = {
@@ -217,8 +207,8 @@ class BackendAPI:
     def create_iot_device_if_not_exists(self, address: str):
         """Create iot device if not exists"""
         exists = self.iot_device_exists(address)
-
         log.debug("Iot device exists %s ...", exists)
+        self.ready_api()
 
         if exists.exists:
             return exists
@@ -227,6 +217,7 @@ class BackendAPI:
 
     async def last_iot_device_data_submission(self, iot_device_id: str) -> int:
         """Get last iot device data submission timestamp"""
+        self.ready_api()
 
         url = f"{self.app_host}/api/iot-devices/{iot_device_id}/last-data-submission"
         res = requests.get(